fix(ci): downgrade base docker image to build
We recently discovered that the usage of ubuntu 22.04 requires to have a
recent version of glibc. That's the docker image used to build RIK which
makes binaries unable to run on some systems. This commit downgrades
base docker images to ubuntu 20.04 in order to have an older version of
glibc, and so have a wider compatibility.

Signed-off-by: AlexandreBrg <burgoni@pm.me>